Video control

+++ controllers/videoController.js
@@ -35,7 +35,10 @@ export const postUpload = async (req, res) => {
     fileUrl: path,
     title,
     description,
+    creator: req.user.id
   });
+  req.user.videos.push(newVideo.id);
+  req.user.save()
   res.redirect(routes.videoDetail(newVideo.id));
 };
 export const videoDetail = async (req, res) => {
@@ -43,7 +46,7 @@ export const videoDetail = async (req, res) => {
     params: { id },
   } = req;
   try {
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate("creator");
     res.render("videoDetail", { pageTitle: video.title, video });
   } catch (error) {
     console.log(error);
@@ -56,7 +59,11 @@ export const getEditVideo = async (req, res) => {
   } = req;
   try {
     const video = await Video.findById(id);
-    res.render("EditVideo", { pageTitle: `Edit ${video.title}`, video });
+    if(`${video.creator}` !==`${ req.user.id}`){
+      throw Error();
+    } else {
+      res.render("EditVideo", { pageTitle: `Edit ${video.title}`, video });
+    }
   } catch (error) {
     res.redirect(routes.home);
   }
@@ -79,7 +86,12 @@ export const deleteVideo = async (req, res) => {
     params: { id },
   } = req;
   try {
-    await Video.findByIdAndRemove({ _id: id });
+    const video = await Video.findById(id);
+    if(`${video.creator}` !==`${ req.user.id}`){
+      throw Error();
+    } else {
+      await Video.findByIdAndRemove({ _id: id });
+    }
   } catch (error) {
     console.log(error);
   }
